refactor(core): 3.5/5 change project structure
