+++ lib/bootstrap-local.js
@@ -58,6 +58,26 @@ Module._resolveFilename = function (request, parent) {
     if (request in packages) {
         return packages[request].main;
     } else {
-        return oldResolveFilename.call(this, request, parent);
+        // TODO: handle case when thare a.json file and a .ts file with the same name
+        if (Module.builtinModules.includes(request)) {
+            return oldResolveFilename.call(this, request, parent);
+        } else {
+            let resolved = null;
+            try {
+                resolved = oldResolveFilename.call(this, request, parent);
+            } catch (e) {
+                throw e;
+            }
+
+            if (resolved.match(/[\\\/]node_modules[\\\/]/)) {
+                return resolved;
+            } else if (request.endsWith('.json')) {
+                return resolved;
+            } else {
+                const maybeTs = resolved.replace('.json', '.ts');
+                if (fs.existsSync(maybeTs)) return maybeTs;
+                return resolved;
+            }
+        }
     }
 };
