+++ scripts/validate-commit.ts
@@ -1,5 +1,5 @@
 import { Logger } from "../tools/logger";
-import { CommitMsgValidator } from "../tools/process/git/commit-msg-validate";
+import { CommitMsgValidator } from "../tools/process/git/commit-msg";
 
 export async function main(args: any, root: string, logger: Logger): Promise<number> {
     const engine = new CommitMsgValidator(process.env["HUSKY_GIT_PARAMS"], logger);
